server/requirements.txt
# ==============================================================================
# EDVANTA BACKEND - REQUIREMENTS.txt (VERCEL OPTIMIZED)
# ==============================================================================
# Generated: 2025-11-28
# Python: 3.10+
# Optimized for Vercel Serverless deployment (under 250MB limit)
# All features including video generation now work on Vercel!
# ==============================================================================

# ------------------------------------------------------------------------------
# CORE WEB FRAMEWORK
# ------------------------------------------------------------------------------
Flask==3.1.1                    # Web framework
flask-cors==5.0.0               # CORS support for API endpoints

# ------------------------------------------------------------------------------
# ENVIRONMENT & CONFIGURATION
# ------------------------------------------------------------------------------
python-dotenv>=1.0.0            # Load environment variables from .env files

# ------------------------------------------------------------------------------
# HTTP & NETWORKING
# ------------------------------------------------------------------------------
requests>=2.31.0                # HTTP requests for external APIs

# ------------------------------------------------------------------------------
# DATABASE
# ------------------------------------------------------------------------------
pymongo==4.6.1                  # MongoDB driver and BSON support

# ------------------------------------------------------------------------------
# CLOUD SERVICES
# ------------------------------------------------------------------------------
cloudinary>=1.34.0              # File/video uploads to Cloudinary

# ------------------------------------------------------------------------------
# GOOGLE GEMINI AI
# ------------------------------------------------------------------------------
google-generativeai>=0.8.0      # Google Gemini API client
google-auth>=2.22.0             # Google authentication utilities

# ------------------------------------------------------------------------------
# PDF & DOCUMENT PROCESSING
# ------------------------------------------------------------------------------
pypdf>=4.0.0                    # PDF reading and text extraction (modernized PyPDF2)
python-docx>=1.0.0              # Microsoft Word document processing
reportlab>=4.2.0                # PDF generation (roadmaps, resumes)

# ------------------------------------------------------------------------------
# IMAGE PROCESSING (Lightweight)
# ------------------------------------------------------------------------------
Pillow>=10.0.0                  # Image manipulation (PIL)

# ==============================================================================
# VIDEO GENERATION NOW USES EXTERNAL APIs (No heavy dependencies needed!)
# - Creatomate API (Free tier: 25 video renders/month)
# - Text-to-Speech: Gemini AI built-in or ElevenLabs free tier
# ==============================================================================


